Add validations and change field types

+++ RetailStore/Aplication/Response/SingleSaleResponse.cs
@@ -3,11 +3,11 @@
     public class SingleSaleResponse
     {
         public int id {  get; set; }
-        public decimal totalPay { get; set;}
+        public double totalPay { get; set;}
         public int totalQuantity { get; set; }
         public decimal totalDiscount { get; set; }
-        public decimal subtotal {  get; set; }
-        public decimal taxes { get; set; }
+        public double subtotal {  get; set; }
+        public double taxes { get; set; }
         public DateTime date { get; set; }
         public List<SingleSaleProduct> products { get; set; }
     }
@@ -16,7 +16,7 @@
     {
         public Guid productId { get; set;}
         public int quantity {  get; set; }
-        public decimal price { get; set; }
+        public double price { get; set; }
         public int discount {  get; set; }
 
     }
